change data type

+++ deploy-api/main.py
@@ -2,8 +2,8 @@ from fastapi import FastAPI
 from pydantic import BaseModel
 
 class Data(BaseModel):
-  x: float
-  y: float
+  x: int
+  y: int
 
 app = FastAPI()
